Feature: Create edci-Certificate
referencing issue #4:
Added BCert Class, that handles the openBadge to edci and edci to openBadge conversion.
An BCert Object can be created using the named Constructors BCert::from_edci($edci) and BCert::from_json($json). The methods get_ob_cert() and get_edci_cert() yield the converted results. using the method add_institution_token($salt) a $salt can be added to an existing certificate.

The edci cert is created shortly after a hash of the openBadge cert was added to the blockchain in the method to_blockchain in locallib.php. The edci cert is then added to the $issuedCertificate object just like the openBadge cert just before the $issuedCertificate is saved to the database.

Todo: needs testing with different certificates
Todo: edci certificate needs to be attached to pdf