Automated certificate signing #14
A course manager can now set a certifier and his or her private key. That private key is stored in the database after getting encrypted with a key stored somewhere in  moodles filesystem. During the automated process, certificates are signed by the the set certifier and his or her private key.
A third party module was added to composer.json. PHP-Encrypt is prvides an api to savely encrypt and decrypt data meant to be kept a secret.

@riegerj TODO: Could you test, it this works on the dev system