Bug fix: Certificates with empty tags and criteria now validate
Bug: Older certificates contained empty tags and criteria, newer certificates did'nt include these attributes when there weren't any tags and criteria set. This caused problems when trying to validate older certificates that were downloaded after the update, that changed the certificate creation process.

Fix: Empty tags and criteria now don't count as unset, so that these attribute will not disappear when creating a new certificate of an older database record. Newer certificates that don't contain these attributes should not be affected by this change. 

Downloaded certificates, that are affected by this bug need to be downloaded again. After that they should validate properly.